+++ include/GameObject.h
@@ -7,25 +7,25 @@ class GameObject {
 
   public:
     // Constructors
-    GameObject(int x, int y, int width, int height, SDL_Color color)
+    GameObject(double x, double y, double width, double height, SDL_Color color)
         : _x(x), _y(y), _width(width), _height(height), _color(color){};
 
-    GameObject(int x, int y, double radius, SDL_Color color)
+    GameObject(double x, double y, double radius, SDL_Color color)
         : _x(x), _y(y), _radius(radius), _color(color){};
 
     // Getters
-    int getX() const;
-    int getY() const;
-    int getWidth() const;
-    int getHeight() const;
+    double getX() const;
+    double getY() const;
+    double getWidth() const;
+    double getHeight() const;
     double getRadius() const;
     SDL_Color getColor() const;
 
     // Setters
-    void setX(int x);
-    void setY(int y);
-    void setWidth(int width);
-    void setHeight(int height);
+    void setX(double x);
+    void setY(double y);
+    void setWidth(double width);
+    void setHeight(double height);
     void setRadius(double radius);
     void setColor(SDL_Color color);
 
@@ -33,10 +33,10 @@ class GameObject {
     virtual void draw(SDL_Renderer *renderer) = 0;
 
   private:
-    int _x;
-    int _y;
-    int _width;
-    int _height;
+    double _x;
+    double _y;
+    double _width;
+    double _height;
     double _radius;
     SDL_Color _color;
 };
