+++ src/Brick.cpp
@@ -11,6 +11,7 @@ void Brick::setContainsBall(int containsBall) { _containsBall = containsBall; }
 void Brick::draw(SDL_Renderer *renderer) {
     SDL_SetRenderDrawColor(renderer, getColor().r, getColor().g, getColor().b,
                            getColor().a);
-    SDL_Rect rect = {getX(), getY(), getWidth(), getHeight()};
+    SDL_Rect rect = {(int)getX(), (int)getY(), (int)getWidth(),
+                     (int)getHeight()};
     SDL_RenderFillRect(renderer, &rect);
 }
