+++ src/GameObject.cpp
@@ -1,24 +1,24 @@
 #include "GameObject.h"
 
-int GameObject::getX() const { return _x; }
+double GameObject::getX() const { return _x; }
 
-int GameObject::getY() const { return _y; }
+double GameObject::getY() const { return _y; }
 
-int GameObject::getWidth() const { return _width; }
+double GameObject::getWidth() const { return _width; }
 
-int GameObject::getHeight() const { return _height; }
+double GameObject::getHeight() const { return _height; }
 
 double GameObject::getRadius() const { return _radius; }
 
 SDL_Color GameObject::getColor() const { return _color; }
 
-void GameObject::setX(int x) { _x = x; }
+void GameObject::setX(double x) { _x = x; }
 
-void GameObject::setY(int y) { _y = y; }
+void GameObject::setY(double y) { _y = y; }
 
-void GameObject::setWidth(int width) { _width = width; }
+void GameObject::setWidth(double width) { _width = width; }
 
-void GameObject::setHeight(int height) { _height = height; }
+void GameObject::setHeight(double height) { _height = height; }
 
 void GameObject::setRadius(double radius) { _radius = radius; }
